+++ src/Components/Progress.js
@@ -6,13 +6,13 @@ class Progress extends Component{
   }
 
   render(){
-    return (
-    <section class="progress">
-      <h2>
-        Raised <span class="secondary">$0</span> of
-        <span class="secondary">$1000</span>
-      </h2>
-    </section>
+    return(
+      <section class="progress">
+        <h2>
+          Raised <span class="secondary">$0</span> of
+          <span class="secondary">$1000</span>
+        </h2>
+      </section>
     )
   }
 }
