+++ src/Components/RecentDonations.js
@@ -1,4 +1,5 @@
 import { Component } from "react";
+import donationData from "../data/donationData";
 
 class RecentDonations extends Component{
   constructor(){
@@ -6,13 +7,17 @@ class RecentDonations extends Component{
   }
 
   render(){
+    let { title } = this.props;
+    let recently = donations.map((donation)=>{
+      return <donationData name={donation.name} amount={donation.amount} caption={donation.caption} />
+    })
+
     return (
       <section>
         <h2>Recent Donations</h2>
         <ul>
-          <li><span>Jo donated $25</span>You really need this. Really.</li>
-          <li><span>Rami donated $10</span>Here, take a break from work!</li>
-          {/* <!-- etc... --> */}
+          <li><span>{ this.name } donated ${ this.amount }</span>{ this.caption }</li>
+          <li><span>{ this.name } donated ${ this.amount }</span>{ this.caption }</li>
         </ul>
       </section>
     );
